refactor: board data GET from CacheableBoardReadStorage
